+++ naga/src/back/msl/mod.rs
@@ -540,7 +540,7 @@ impl Options {
                     }
                     // macOS: Since Metal 2.2
                     // iOS: Since Metal 2.3 (check depends on https://github.com/gfx-rs/wgpu/issues/4414)
-                    crate::BuiltIn::Barycentric if self.lang_version < (2, 3) => {
+                    crate::BuiltIn::Barycentric { .. } if self.lang_version < (2, 3) => {
                         return Err(Error::UnsupportedAttribute("barycentric_coord".to_string()));
                     }
                     _ => {}
@@ -695,7 +695,10 @@ impl ResolvedBinding {
                     Bi::PointCoord => "point_coord",
                     Bi::FrontFacing => "front_facing",
                     Bi::PrimitiveIndex => "primitive_id",
-                    Bi::Barycentric => "barycentric_coord",
+                    Bi::Barycentric { perspective: true } => "barycentric_coord",
+                    Bi::Barycentric { perspective: false } => {
+                        "barycentric_coord, center_no_perspective"
+                    }
                     Bi::SampleIndex => "sample_id",
                     Bi::SampleMask => "sample_mask",
                     // compute