+++ naga/src/front/wgsl/parse/conv.rs
@@ -50,7 +50,8 @@ pub fn map_built_in(
         "front_facing" => crate::BuiltIn::FrontFacing,
         "frag_depth" => crate::BuiltIn::FragDepth,
         "primitive_index" => crate::BuiltIn::PrimitiveIndex,
-        "barycentric" => crate::BuiltIn::Barycentric,
+        "barycentric" => crate::BuiltIn::Barycentric { perspective: true },
+        "barycentric_no_perspective" => crate::BuiltIn::Barycentric { perspective: false },
         "sample_index" => crate::BuiltIn::SampleIndex,
         "sample_mask" => crate::BuiltIn::SampleMask,
         // compute
@@ -111,6 +112,7 @@ pub fn map_interpolation(word: &str, span: Span) -> Result<'_, crate::Interpolat
         "linear" => Ok(crate::Interpolation::Linear),
         "flat" => Ok(crate::Interpolation::Flat),
         "perspective" => Ok(crate::Interpolation::Perspective),
+        "per_vertex" => Ok(crate::Interpolation::PerVertex),
         _ => Err(Box::new(Error::UnknownAttribute(span))),
     }
 }