+++ wgpu-hal/src/vulkan/adapter.rs
@@ -563,7 +563,9 @@ impl PhysicalDeviceFeatures {
             shader_barycentrics: if enabled_extensions
                 .contains(&khr::fragment_shader_barycentric::NAME)
             {
-                let needed = requested_features.intersects(wgt::Features::SHADER_BARYCENTRICS);
+                let needed = requested_features.intersects(
+                    wgt::Features::SHADER_BARYCENTRICS | wgt::Features::SHADER_PER_VERTEX,
+                );
                 Some(
                     vk::PhysicalDeviceFragmentShaderBarycentricFeaturesKHR::default()
                         .fragment_shader_barycentric(needed),
@@ -745,7 +747,7 @@ impl PhysicalDeviceFeatures {
 
         if let Some(ref shader_barycentrics) = self.shader_barycentrics {
             features.set(
-                F::SHADER_BARYCENTRICS,
+                F::SHADER_BARYCENTRICS | F::SHADER_PER_VERTEX,
                 shader_barycentrics.fragment_shader_barycentric != 0,
             );
         }
@@ -1292,8 +1294,11 @@ impl PhysicalDeviceProperties {
             extensions.push(ext::mesh_shader::NAME);
         }
 
-        // Require `VK_KHR_fragment_shader_barycentric` if the associated feature was requested
-        if requested_features.intersects(wgt::Features::SHADER_BARYCENTRICS) {
+        // Require `VK_KHR_fragment_shader_barycentric` if an associated feature was requested
+        // Vulkan bundles both barycentrics and per-vertex attributes under the same feature.
+        if requested_features
+            .intersects(wgt::Features::SHADER_BARYCENTRICS | wgt::Features::SHADER_PER_VERTEX)
+        {
             extensions.push(khr::fragment_shader_barycentric::NAME);
         }
 
@@ -2341,7 +2346,10 @@ impl super::Adapter {
                 capabilities.push(spv::Capability::ClipDistance);
             }
 
-            if features.intersects(wgt::Features::SHADER_BARYCENTRICS) {
+            // Vulkan bundles both barycentrics and per-vertex attributes under the same feature.
+            if features
+                .intersects(wgt::Features::SHADER_BARYCENTRICS | wgt::Features::SHADER_PER_VERTEX)
+            {
                 capabilities.push(spv::Capability::FragmentBarycentricKHR);
             }